Added major database functions.
The game will now pull items from a JSON file into the game item database
at startup. Items can be created with a constructor, and there is an erase
constructor that sets a slot to an empty item. Added method to search
through the database by item ID, this will return the item if any ID has
matched. Added 2 items for testing purposes.